requirements.txt
django==2.0.4
mongoengine
numpy
netmiko
jinja2
netaddr
json
napalm
ntplib
django-rest-framework-mongoengine
django-background-tasks
djangorestframework
markdown
django-filter
